Add settings.admin check on commands
Forgot that before I began working on the botlib base separately from my
first few bots that I made this check manually in the commands that
utilized it. Now multiple ids can be added to the settings.admin array.
This isn't really necessary since the intent of this settings property is
for commands that are meant to be run only by bot operators; people who
have access to the bots code itself, which is generally only one person.
Commands that are meant to be for server admins should utilize the
"ADMINISTRATOR" permission in the Command#permissions property.