Dereference stored objects on get/set
It took a lot of frustrating trial and error to realize what the problem was, and when I did realize it, I forgot to fix it for more than just adding default settings to guilds. This is my second time dealing with issues surrounding how javascript passes many things by reference rather than by value, so it's embarassing that it took so long to narrow it down.