Add command metadata decorators
Can be used in place of a constructor with a CommandInfo object when creating commands with TypeScript. Not sure if these decorators are Babel compatible or not. The bare minimum metadata a command needs to function is @name, @description, and @usage. I will likely apply this to CommandInfo as well later on.