Remove unnecessary client errors, let client be created without token
The client will not be able to be started, but this allows me to use the Client class in some scripts that don't need discord-specific stuff without the need for connecting to discord.

This commit also allows clients to be created without passing in a commandsDir. Also useful for the scripts thing.