Add Logger#removeBaseTransport, update docs
I have an idea for something potentially more desirable than this solution but I'm letting it sit for a while to see how I feel about it. I like the base transport anyway since It was what I originally wrote for the logger based on how I liked my logging to look, so I personally won't be removing the base transport. I'll need other people to weigh in on it