Clean up CommandDispatcher
It was in dire need. Overhauling it during the TypeScript port was hasty due to wanting to hurry and finish porting everything else and it ended up pretty sloppy. I think I've really ironed it out now. Everything feels much easier to read and comprehend, and I simplified quite a few things